Suppress the git main branch name message on bup init
If the git config option init.defaultBranch is not set, git init prints a
message on how to change the default branch name. The default branch name is
never used, so the message is irrelevant and might be confusing for users.
Setting some explicit default branch name suppresses that warning.

Signed-off-by: Moritz Lell <mlell08@gmail.com>